// apps/web/app/routes/dokument/types/api/JournalpostTypes.ts
export type Sakstilknytning = string;
// ...
export class EndreJournalpostRequest {
    skalJournalfores = false;
    journalpostId: string;
    journalforendeEnhet?: string;
    behandlingstema?: string;
    beskrivelse?: string;
    dokumentDato?: string;
    gjelderType?: string;
    fagomrade?: string;
    journaldato?: string;
    gjelder?: string;
    tittel?: string;
    tilknyttSaker?: Sakstilknytning[];
    avsenderNavn?: string;
    endreDokumenter?: EndreDokument[];
    endreReturDetaljer?: EndreReturDetaljer[];

    constructor(journalpostId: string) {
        this.journalpostId = journalpostId;
    }
}

export class LagreJournalpostRequest extends EndreJournalpostRequest {
    override skalJournalfores = false;
}

export class RegistrerJournalpostRequest extends EndreJournalpostRequest {
    override skalJournalfores = true;
}

export interface JournalpostToRegister {
    journalpostId: string;
    journalforendeEnhet: string;
    mottatDato?: string;
    endreDokumenter: EndreDokument[];
    tilknyttSaker?: string[];
    tittel?: string;
    gjelderIdent: string;
    avsenderNavn?: string;
}

export interface EndreDokument {
    dokId: string;
    tittel: string;
}

export interface EndreReturDetaljer {
    originalDato: string;
    nyDato: string;
    beskrivelse: string;
}